requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
streamlit==1.36.0
python-dotenv==1.0.1
openai>=1.99.1,<2.0.0
chromadb==0.5.3
pypdf==4.3.1
pydantic>=2.10.0
requests==2.32.3
tqdm==4.66.4
httpx<0.28
